requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Computer Vision & AI
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
deepface==0.0.81
ultralytics==8.0.196

# Face Detection & Tracking
yolov5==7.0.13
bytetracker==0.3.2

# Data Science
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4

# Database
sqlite3
# Note: sqlite3 is included with Python standard library

# Communication
twilio==8.12.0

# Utilities
Pillow==10.1.0
requests==2.31.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Logging
logging
# Note: logging is included with Python standard library

# Environment
python-dotenv==1.0.0

# Additional ML/AI dependencies
tensorflow==2.15.0
# Note: tensorflow is optional, remove if not using TensorFlow models

# Image processing
opencv-contrib-python==4.8.1.78
# Use this instead of opencv-python if you need additional features

# Threading and async
asyncio
# Note: asyncio is included with Python standard library